refactor: Rename provider to providers
